# db: enforce batch point visibility in the merging iterator