@erikgrinaker 2022-09-07
Awesome, thanks @jbowens! I'll leave a proper review to the storage team, but this seems great.